IKLOCSEN-1744 Display rotation algo not working on Surnia/Otus L
With the upgrade of the sensor hub hal on L, the display rotation algo is broken on Surnia/Otus.
User does not notice as the AP based display rotation is kicked in, however this needs to be fixed.

Change-Id: I005a2689964d72b2226118165b5f4f85de0b53ef
Reviewed-on: http://gerrit.mot.com/680393
Tested-by: Jira Key <jirakey@motorola.com>
Reviewed-by: Manasa Guntaka <mguntaka@motorola.com>
Reviewed-by: Robert Wiegand <Robert.P.Wiegand@motorola.com>
SLTApproved: Iyengar Sriram Kaushik <sriramkaushik@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>